% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ook-package.R
\docType{package}
\name{ook-package}
\alias{ook}
\alias{ook-package}
\title{ook: Install commonly used packages for a new R user. Ook!}
\description{
ook? installs commonly used packages via the packages it Imports.
Primarily aimed at a user new to R. Installing ook will install
a bunch of R packages that will help a new R user get up and running. Ook...
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/adam-gruer/ook}
  \item Report bugs at \url{https://github.com/adam-gruer/ook/issues}
}

}
\author{
\strong{Maintainer}: Adam Gruer \email{adamgruer@icloud.com}

}
\keyword{internal}
